# alewlsh/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alewlsh Emergency Response API",
    "description": "API documentation for the Alewlsh emergency response and safety platform. This API provides endpoints for user authentication, panic alerts, emergency contacts, safety features, and police coordination.",
    "version": "1.0.0",
    "contact": {
      "name": "Alewlsh Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"]
      },
...
      "EmergencyContact": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "relationship": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["name", "phone"]
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user with phone number and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "User phone number",
                    "example": "0911111111"
                  },
                  "password": {
                    "type": "string",
                    "description": "User password",
                    "example": "testpassword"
                  }
                },
                "required": ["phone", "password"],
                "example": {
                  "phone": "0911111111",
                  "password": "testpassword"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/emergency-contacts": {
      "get": {
        "tags": ["Emergency Contacts"],
        "summary": "Get emergency contacts",
        "description": "Retrieve all emergency contacts for the authenticated user",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Emergency contacts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contacts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EmergencyContact"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Emergency Contacts"],
        "summary": "Add emergency contact",
        "description": "Add a new emergency contact for the authenticated user",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact name"
                  },
                  "phone": {
                    "type": "string",
                    "description": "Contact phone number"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Contact email (optional)"
                  },
                  "relationship": {
                    "type": "string",
                    "description": "Relationship to user (optional)"
                  }
                },
                "required": ["name", "phone"],
                "example": {
                  "name": "Emergency Contact",
                  "phone": "0911111111",
                  "email": "contact@example.com",
                  "relationship": "friend"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Emergency contact added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "contact": {
                      "$ref": "#/components/schemas/EmergencyContact"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}